refactor: rename Checkr loading component, enhance Checkr page with user contract fetching, and update AppSidebar to use Suspense
