+++ src/app/loading.tsx
@@ -1,0 +1,5 @@
+"use client";
+
+export default function Loading() {
+  return <div>Loading...</div>;
+}
