+++ src/components/chat/use-chat-id.ts
@@ -1,3 +1,5 @@
+"use client";
+
 import { parseAsString, useQueryState } from "nuqs";
 
 export function useChatId() {
