+++ src/components/chat/use-chat-open.ts
@@ -1,3 +1,5 @@
+"use client";
+
 import { parseAsBoolean, useQueryState } from "nuqs";
 
 export function useChatOpen() {
